refactor: use non-deprecated indexmap methods
The `remove()` and `take()` methods are deprecated in favor of
`swap_remove()`, `shift_remove()`, `swap_take()`, and `shift_take()`,
which have been available since indexmap 1.2.0.

The "shift" variants are used in StGit because they will preserve set
order at the cost of being O(n). But since parent and patch sets are not
particularly large, perhaps not a big concern.