fix: stg import --reject should create empty commit
When trying to import a patch that does not apply cleanly, and using
--reject option, stg should apply what it can, leave the rest in .rej
files, and create an empty commit. The work to apply the patch is
outsourced to "git apply --reject" which exits with status code 1 if
patch is applied partially and it is treated as error by stg import.

Fix the issue by treating exit codes less than 128 from "git apply
--reject" as success (128 signals hard error).

Also add a test case and fix up documentation for "stg import".

Closes: #471

Signed-off-by: Dmitry Torokhov <dtor@google.com>