fix(push): report modified patches in output
Somewhere along the way, `stg push` stopped correctly reporting to the
user when the pushed patch(es) were updated during the push.

With this change, when a push causes a new commit to be made for a
patch, that patch will be reported in the user-facing output as
"(modified)" unless a superceding "(empty)", "(merged)", or "(conflict)"
status applies.

Fixes: #464