fix(squash): preserve consensus author
When all of the patches to be squashed have the same author, use that
author for the squashed commit.

If the patches to squash have heterogeneous authors, default to the
configured user for the repo. (This was the behavior for all squashes.)

The command line flags (--author, --authname, --authemail) still take
precedence.

Fixes: #490