fix(squash): keep author when squashed commits have same author
When the commits being squashed all share the same author, the resulting
squashed commit should have that same author.

This was the intent of the code, but it was broken due to author
comparisons comparing entire signatures, which includes the author
**time**, instead of just comparing the author name and email. I.e. if
the author times were different, then the authors were deemed different
even if they shared the same name and email.

The comparisons using time enabled automated tests to not see this issue
because in automated scenarios patches could all be created quickly
enough to share the same timestamp.

Fixes: #592